requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
chromadb
pypdf
python-dotenv